typecasts
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_0branch/bzflag@12365 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ plugins/nagware/nagware.cpp
@@ -121,7 +121,7 @@ void sendNagMessage (int who, std::string *msg ){
   std::string fullMsg = *msg + Config.msgSuffix;
   unsigned int idx=0, x;
   
-  while ((x = fullMsg.find("\\n", idx)) != std::string::npos){
+  while ((x = fullMsg.find("\\n", idx)) != (unsigned int)std::string::npos){
     bz_sendTextMessage(BZ_SERVER, who, fullMsg.substr(idx, x-idx).c_str());
     idx = x+2;
   }
@@ -322,7 +322,7 @@ void Nagware::process ( bz_EventData *eventData )
   // tick
   } else if (eventData->eventType == bz_eTickEvent) {
     bz_TickEventData *msgData = (bz_TickEventData*)eventData;
-    tickEvent (msgData->time);
+    tickEvent ((float)msgData->time);
 
   }
 }
